requirements.txt
Flask==3.0.0
qrcode==7.4
pillow==9.4.0